+++ docs/tutorials/pisa/pisa.Person-1.pdf
@@ -70,7 +70,7 @@ endobj
 % 'R7': class PDFInfo 
 7 0 obj
 << /Author ()
- /CreationDate (D:20150904151728-02'00')
+ /CreationDate (D:20150905053556-02'00')
  /Creator (\(unspecified\))
  /Keywords ()
  /Producer (pisa HTML to PDF <http://www.htmltopdf.org>)
@@ -124,7 +124,7 @@ xref
 trailer
 << /ID 
  % ReportLab generated PDF document -- digest (http://www.reportlab.com) 
- [(1&G`=\221\246A\026\303=\373\231\002uL) (1&G`=\221\246A\026\303=\373\231\002uL)] 
+ [(\366\337\276_m\204\316\262\377\263\035P\335\265\213\323) (\366\337\276_m\204\316\262\377\263\035P\335\265\213\323)] 
 
  /Info 7 0 R
  /Root 6 0 R
