+++ lino/modlib/bootstrap3/config/bootstrap3/detail.html
@@ -4,7 +4,8 @@
 {% block main %}
 <div class="row-fluid">
   <h1>{{heading}}</h1>
-  {{E.tostring(main)}}
+  {{E.tostring(navigator)}}
+  {{E.tostring(form)}}
 </div>
 {% endblock %}
